+++ webapp/deck/forms.py
@@ -1,0 +1,8 @@
+from flask_wtf import FlaskForm
+from wtforms import StringField, SubmitField
+
+
+class DeckForm(FlaskForm):
+    name = StringField("Название колоды", validators=[DataRequired()], render_kw={"class": "form-control"})
+    comment = StringField("Комментарий колоды", render_kw={"class": "form-control"})
+    button = SubmitField("Сохранить", render_kw={"class": "btn btn-primary"})
