+++ webapp/templates/base.html
@@ -21,8 +21,8 @@
                 Колоды
               </a>
               <ul class="dropdown-menu">
-                <li><a class="dropdown-item" href='{{ url_for("decks_view") }}'>Список колод</a></li>
-                <li><a class="dropdown-item" href='{{ url_for("deck_new") }}'>Создать колоду</a></li>
+                <li><a class="dropdown-item" href='{{ url_for("deck.decks_view") }}'>Список колод</a></li>
+                <li><a class="dropdown-item" href='{{ url_for("deck.deck_new") }}'>Создать колоду</a></li>
               </ul>
             </li>
             <li class="nav-item">
@@ -36,9 +36,9 @@
           <ul class="navbar-nav mb-2 mb-lg-0">
             <li class="nav-item">
               {% if current_user.is_authenticated %}
-              <a class="nav-link active" aria-current="page" href='{{ url_for("logout") }}'>Logout</a>
+              <a class="nav-link active" aria-current="page" href='{{ url_for("user.logout") }}'>Logout</a>
               {% else %}            
-              <a class="nav-link active" aria-current="page" href='{{ url_for("login") }}'>Login</a>
+              <a class="nav-link active" aria-current="page" href='{{ url_for("user.login") }}'>Login</a>
               {% endif %}
             </li>
           </ul>
