+++ webapp/templates/card/add_new_card_form.html
@@ -9,33 +9,33 @@
         <p>
              {{ card_form.deck.label }}
              {{ card_form.deck() }}
-             <a href={{ url_for("deck_new") }} >Добавить колоду</a>
+             <a href='{{ url_for("deck.deck_new") }}' >Добавить колоду</a>
         </p>
-    <p>
-        {{ card_form.side_1.label }}
-        {{ card_form.side_1() }}
-    </p>
-   <p>        
-        {{ card_form.side_2.label }}
-        {{ card_form.side_2() }}
-    </p>
-   <p> 
-        {{ card_form.type.label }}
-        {{ card_form.type() }} 
-    </p>
-    <p>
-        {{ card_form.tags.label }}
-        {{ card_form.tags() }}   
-    </p>
-   <p>
-       {{ card_form.is_active() }}   
-       {{ card_form.is_active.label }}
-   </p>
+        <p>
+            {{ card_form.side_1.label }}
+            {{ card_form.side_1() }}
+        </p>
+        <p>
+            {{ card_form.side_2.label }}
+            {{ card_form.side_2() }}
+        </p>
+        <p>
+            {{ card_form.type.label }}
+            {{ card_form.type() }}
+        </p>
+        <p>
+            {{ card_form.tags.label }}
+            {{ card_form.tags() }}
+        </p>
+        <p>
+           {{ card_form.is_active() }}
+           {{ card_form.is_active.label }}
+        </p>
 
-    <p>
-        {{ card_form.button() }} 
-    </p>
-
+        <p>
+            {{ card_form.button() }}
+        </p>
+    </form>
     
 </div>
