+++ webapp/templates/card/edit_card_form.html
@@ -6,32 +6,32 @@
     <form action="" method="POST">
         {{ card_form.hidden_tag() }}
         
-    <p>
-        {{ card_form.side_1.label }}
-        {{ card_form.side_1() }}
-    </p>
-   <p>        
-        {{ card_form.side_2.label }}
-        {{ card_form.side_2() }}
-    </p>
-   <p> 
-        {{ card_form.type.label }}
-        {{ card_form.type() }} 
-    </p>
-    <p>
-        {{ card_form.tags.label }}
-        {{ card_form.tags() }}   
-    </p>
-   <p>
-       {{ card_form.is_active() }}   
-       {{ card_form.is_active.label }}
-   </p>
+        <p>
+            {{ card_form.side_1.label }}
+            {{ card_form.side_1() }}
+        </p>
+        <p>
+            {{ card_form.side_2.label }}
+            {{ card_form.side_2() }}
+        </p>
+        <p>
+            {{ card_form.type.label }}
+            {{ card_form.type() }}
+        </p>
+        <p>
+            {{ card_form.tags.label }}
+            {{ card_form.tags() }}
+        </p>
+        <p>
+           {{ card_form.is_active() }}
+           {{ card_form.is_active.label }}
+        </p>
 
-    <p>
-        {{ card_form.button() }} 
-    </p>
+        <p>
+            {{ card_form.button() }}
+        </p>
 
-    
+    </form>
 </div>
