+++ webapp/templates/deck/deck_with_cards.html
@@ -14,7 +14,7 @@
         <h5 class="card-title">{{ deck.name }}</h5>
         <p class="card-text">{{ deck.comment }}</p>
         <div class="btn-group">
-            <button type="button" class="btn btn-primary"><a class="dropdown-item" href={{ url_for("deck_study", deck_id=deck.id )}}>Повторять карточки</a></button>
+            <button type="button" class="btn btn-primary"><a class="dropdown-item" href='{{ url_for("deck.deck_study", deck_id=deck.id )}}'>Повторять карточки</a></button>
             <button type="button" class="btn btn-primary dropdown-toggle dropdown-toggle-split" data-bs-toggle="dropdown"
                 aria-expanded="false">
                 <span class="visually-hidden">Toggle Dropdown</span>
@@ -44,7 +44,7 @@
 <!--cards-->
 <div class="list-group">
     {% for card in deck.card: %}
-    <a href={{ url_for("edit_card", card_id=card.id) }} class="list-group-item list-group-item-action">
+    <a href='{{ url_for("card.edit_card", card_id=card.id) }}' class="list-group-item list-group-item-action">
         <div class=" d-flex w-100 justify-content-between">
             <div>
                 <h5 class="mb-1">{{card.side_1}}</h5>
