+++ webapp/templates/deck/decks_view.html
@@ -2,11 +2,11 @@
 
 {% block content %}
 
-<div class="d-flex justify-content-end""> 
+<div class="d-flex justify-content-end">
 </div>
     <div class="list-group">
         {% for deck in decks: %}
-        <a href={{ url_for("deck_view", deck_id=deck.id) }} class="list-group-item list-group-item-action"">
+        <a href='{{ url_for("deck.deck_view", deck_id=deck.id) }}' class="list-group-item list-group-item-action">
             <div class="d-flex w-100 justify-content-between">
                 <h4 class="mb-1">{{deck.name}}</h4>
                 <small>{{ deck.card_count }}</small>
