+++ webapp/templates/login.html
@@ -15,7 +15,7 @@
     <p>
         {{ form.button }}
     </p>
-    <a href='{{ url_for("signup") }}'> Зарегистрироваться </a>
+    <a href='{{ url_for("user.signup") }}'> Зарегистрироваться </a>
 </form>
 </div>
