+++ webapp/templates/signup.html
@@ -2,7 +2,7 @@
 
 {% block content %}
 <div class="d-flex justify-content-center">
-    <form action="{{ url_for('process_signup') }}" method="post">
+    <form action="{{ url_for('user.process_signup') }}" method="post">
         {{ form.hidden_tag() }}
         <p>
             {{ form.username.label }}
